refactor: 공통 fetch 분리 및 orval mutator 시그니처 유지만 담당하도록 수정
