+++ src/lib/api/api-client.ts
@@ -1,6 +1,12 @@
 /* eslint-disable no-console */
 import type { ApiErrorResponse } from "@/types/shared/types";
 
+export interface ExecuteFetchOptions {
+  timeout?: number;
+  retry?: RetryOptions;
+  signal?: AbortSignal;
+}
+
 export const isDev = process.env.NODE_ENV === "development";
 export const API_URL = process.env.NEXT_PUBLIC_API_URL;
 
@@ -78,4 +84,95 @@ export function buildRetryConfig(retry?: RetryOptions) {
     retryDelay: retry?.retryDelay ?? 1000,
     retryableStatuses: retry?.retryableStatuses ?? [408, 429, 500, 502, 503, 504],
   };
+}
+
+// ===== 공통 fetch 실행 =====
+
+export async function executeFetch<T>(
+  method: RequestMethod,
+  url: string,
+  init: RequestInit,
+  options?: ExecuteFetchOptions
+): Promise<T> {
+  const retryConfig = buildRetryConfig(options?.retry);
+  const timeout = options?.timeout ?? 30000;
+  let attempt = 0;
+
+  while (true) {
+    const controller = new AbortController();
+    const timeoutId = setTimeout(() => controller.abort(), timeout);
+
+    try {
+      log("request", method, url, init.body);
+
+      const response = await fetch(url, {
+        ...init,
+        signal: options?.signal || controller.signal,
+      });
+
+      clearTimeout(timeoutId);
+
+      if (!response.ok) {
+        let errorData: ApiErrorResponse | null = null;
+        try {
+          errorData = await response.json();
+        } catch {
+          // JSON 파싱 실패 시 무시
+        }
+
+        const error = new ApiError(
+          errorData?.error.message ?? `HTTP error! status: ${response.status}`,
+          response.status,
+          errorData
+        );
+
+        if (shouldRetry(error, attempt, retryConfig.maxRetries, retryConfig.retryableStatuses)) {
+          attempt++;
+          const delay = retryConfig.retryDelay * Math.pow(2, attempt - 1);
+          log("error", `Retry ${attempt}/${retryConfig.maxRetries} after ${delay}ms`, error);
+          await sleep(delay);
+          continue;
+        }
+
+        log("error", response.status, errorData);
+        throw error;
+      }
+
+      if (response.status === 204) {
+        log("response", response.status, "No Content");
+        return null as T;
+      }
+
+      const responseData = await response.json();
+      log("response", response.status, responseData);
+      return responseData;
+    } catch (error) {
+      clearTimeout(timeoutId);
+
+      if (error instanceof Error && error.name === "AbortError") {
+        const timeoutError = new TimeoutError("Request timeout or cancelled");
+        log("error", timeoutError);
+        throw timeoutError;
+      }
+
+      if (error instanceof ApiError) {
+        throw error;
+      }
+
+      const networkError = new NetworkError("Network request failed", error);
+
+      if (
+        shouldRetry(networkError, attempt, retryConfig.maxRetries, retryConfig.retryableStatuses)
+      ) {
+        attempt++;
+        const delay = retryConfig.retryDelay * Math.pow(2, attempt - 1);
+        log("error", `Retry ${attempt}/${retryConfig.maxRetries} after ${delay}ms`, networkError);
+        await sleep(delay);
+        continue;
+      }
+
+      log("error", networkError);
+      throw networkError;
+    }
+  }
 }
