+++ src/lib/api/api.ts
@@ -1,16 +1,4 @@
-import type { ApiErrorResponse } from "@/types/shared/types";
-import {
-  API_URL,
-  ApiError,
-  NetworkError,
-  TimeoutError,
-  type RetryOptions,
-  type RequestMethod,
-  log,
-  sleep,
-  shouldRetry,
-  buildRetryConfig,
-} from "./api-client";
+import { API_URL, type RetryOptions, type RequestMethod, executeFetch } from "./api-client";
 
 interface RequestOptions {
   headers?: Record<string, string>;
@@ -75,96 +63,21 @@ async function request<T>(
     }
   }
 
-  const retryConfig = buildRetryConfig(options?.retry);
-
-  const timeout = options?.timeout ?? 30000;
-  let attempt = 0;
-
-  while (true) {
-    const controller = new AbortController();
-    const timeoutId = setTimeout(() => controller.abort(), timeout);
-
-    try {
-      const config: RequestInit = {
-        method,
-        headers,
-        credentials: "include",
-        signal: options?.signal || controller.signal,
-      };
-
-      if (data && method !== "GET") {
-        config.body = JSON.stringify(data);
-      }
-
-      log("request", method, url.toString(), data);
-
-      const response = await fetch(url.toString(), config);
-
-      clearTimeout(timeoutId);
-
-      if (!response.ok) {
-        let errorData: ApiErrorResponse | null = null;
-        try {
-          errorData = await response.json();
-        } catch {
-          // JSON 파싱 실패 시 무시
-        }
-
-        const error = new ApiError(
-          errorData?.error.message ?? `API Error: ${response.status}`,
-          response.status,
-          errorData
-        );
-
-        if (shouldRetry(error, attempt, retryConfig.maxRetries, retryConfig.retryableStatuses)) {
-          attempt++;
-          const delay = retryConfig.retryDelay * Math.pow(2, attempt - 1);
-          log("error", `Retry ${attempt}/${retryConfig.maxRetries} after ${delay}ms`, error);
-          await sleep(delay);
-          continue;
-        }
-
-        log("error", response.status, errorData);
-        throw error;
-      }
-
-      if (response.status === 204) {
-        log("response", response.status, "No Content");
-        return null as T;
-      }
-
-      const responseData = await response.json();
-      log("response", response.status, responseData);
-      return responseData;
-    } catch (error) {
-      clearTimeout(timeoutId);
-
-      if (error instanceof Error && error.name === "AbortError") {
-        const timeoutError = new TimeoutError("Request timeout or cancelled");
-        log("error", timeoutError);
-        throw timeoutError;
-      }
-
-      if (error instanceof ApiError) {
-        throw error;
-      }
-
-      const networkError = new NetworkError("Network request failed", error);
-
-      if (
-        shouldRetry(networkError, attempt, retryConfig.maxRetries, retryConfig.retryableStatuses)
-      ) {
-        attempt++;
-        const delay = retryConfig.retryDelay * Math.pow(2, attempt - 1);
-        log("error", `Retry ${attempt}/${retryConfig.maxRetries} after ${delay}ms`, networkError);
-        await sleep(delay);
-        continue;
-      }
-
-      log("error", networkError);
-      throw networkError;
+  return executeFetch<T>(
+    method,
+    url.toString(),
+    {
+      method,
+      headers,
+      credentials: "include",
+      body: data && method !== "GET" ? JSON.stringify(data) : undefined,
+    },
+    {
+      timeout: options?.timeout,
+      retry: options?.retry,
+      signal: options?.signal,
     }
-  }
+  );
 }
 
 export const api = {
